refactor: modularize codebase by splitting into separate files
