+++ styles/main.css
@@ -127,6 +127,15 @@ main {
         input {
           width: 100%;
           accent-color: var(--primary-color);
+
+          &:disabled {
+            cursor: not-allowed;
+            opacity: 0.5;
+
+            & + div {
+              opacity: 0.5;
+            }
+          }
         }
 
         div {
@@ -198,6 +207,7 @@ main {
         cursor: grab;
         touch-action: none;
         user-select: none;
+        border: 1px solid #9e9e9e;
 
         &:active {
           cursor: grabbing;
@@ -210,6 +220,10 @@ main {
 
       &[data-shape="circle"] canvas {
         border-radius: 50%;
+      }
+
+      &[data-shape="material"] canvas {
+        border: none;
       }
     }
   }
